# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Najm Educational System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: gradeLevel
          in: query
          schema:
            type: string
        - name: accessRole
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /users/{studentId}:
    get:
      operationId: getUserByStudentId
      tags: [users]
      summary: Get user by student ID (admin only)
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{studentId}/access:
    patch:
      operationId: updateUserAccess
      tags: [users]
      summary: Toggle user access role (admin only)
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccessBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{studentId}/stars:
    patch:
      operationId: updateUserStars
      tags: [users]
      summary: Update user stars balance (admin only)
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStarsBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: string
        email:
          type: string
        fullName:
          type: string
        phone:
          type: string
        gradeLevel:
          type: string
          enum: [grade9, baccalaureate, outcomes, steps1000, ielts]
        accessRole:
          type: string
          enum: [free, paid, admin]
        starsBalance:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - studentId
        - email
        - fullName
        - phone
        - gradeLevel
        - accessRole
        - starsBalance
        - createdAt
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - users
        - total
        - page
        - limit
    UpdateAccessBody:
      type: object
      properties:
        accessRole:
          type: string
          enum: [free, paid, admin]
      required:
        - accessRole
    UpdateStarsBody:
      type: object
      properties:
        starsBalance:
          type: integer
      required:
        - starsBalance
...